% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-rff.R
\name{nn_random_fourier_features}
\alias{nn_random_fourier_features}
\title{Random Fourier Features Layer}
\usage{
nn_random_fourier_features(
  input_dim,
  num_grids = 8,
  dropout = 0,
  activation_expectation = 1.64
)
}
\arguments{
\item{input_dim}{Integer. Input dimension.}

\item{num_grids}{Integer. Number of Fourier frequencies.}

\item{dropout}{Numeric. Dropout probability.}

\item{activation_expectation}{Numeric. Scaling constant for initialization.}
}
\value{
A torch nn_module.
}
\description{
Torch module implementing trainable random Fourier features for
Kolmogorov-Arnold Fourier Networks.
}
