% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.kaf_fit}
\alias{print.kaf_fit}
\title{Print a fitted KAF model}
\usage{
\method{print}{kaf_fit}(x, ...)
}
\arguments{
\item{x}{A fitted object returned by \code{kaf_fit()}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a fitted KAF model
}
