% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkaf-package.R
\docType{package}
\name{rkaf-package}
\alias{rkaf}
\alias{rkaf-package}
\title{rkaf: Kolmogorov-Arnold Fourier Networks in R}
\description{
\code{rkaf} provides an R implementation of Kolmogorov-Arnold Fourier Networks
using the \code{torch} backend.
}
\details{
The package supports:
\itemize{
\item regression
\item binary classification
\item multiclass classification
\item formula and matrix interfaces
\item mini-batch training
\item validation splits
\item early stopping
\item automatic standardization
\item best-model restoration
\item KAF-specific diagnostics
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gsidoine/rkaf}
  \item \url{https://gsidoine.github.io/rkaf/}
  \item Report bugs at \url{https://github.com/gsidoine/rkaf/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Sidoine \email{gjh.sidoine@gmail.com}

}
\keyword{internal}
